requirements/requirements.txt
Django==1.6.2
Pillow==2.4.0
South==1.0
argparse==1.2.1
django-bootstrap3==4.11.0
django-extensions==1.3.3
django-tastypie==0.11.0
gunicorn==19.1.0
path.py==5.1
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.3.0
six==1.6.1
tablib==0.10.0
wsgiref==0.1.2
